+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Added folder support for resources
     - `foo:bar` now supports `foo:foo/foobar`
+- Entity.extinguish() and Entity.extinguishWithSound() are now merged to 1 function Entity.extinguish(boolean)